requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
PyGithub==2.1.1
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.0
tree-sitter==0.20.4
radon==6.0.1
GitPython==3.1.40
aiofiles==23.2.1
